+++ dwds/test/frontend_server_evaluate_test.dart
@@ -413,6 +413,21 @@ void main() async {
                   contains('CompilationError:')));
         });
       });
+
+      test('module load error', () async {
+        await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var error = await setup.service.evaluateInFrame(
+              isolate.id, event.topFrame.index, 'd.deferredPrintLocal()');
+
+          expect(
+              error,
+              isA<ErrorRef>().having((instance) => instance.message, 'message',
+                  contains('LoadModuleError:')));
+        });
+      }, skip: 'https://github.com/dart-lang/sdk/issues/48587');
     });
 
     group('evaluate', () {